Replace size_t with int in a few media classes.
This addressed TODOs in DataSource::Read() and AudioRendererAlgorithmBase, which led to converting Buffer and SeekableBuffer as well.

Review URL: https://chromiumcodereview.appspot.com/9854031

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@129140 0039d316-1c4b-4281-b951-d872f2087c98

+++ media/base/buffers.h
@@ -51,7 +51,7 @@ class MEDIA_EXPORT Buffer : public base::RefCountedThreadSafe<Buffer> {
   virtual const uint8* GetData() const = 0;
 
   // Returns the size of valid data in bytes.
-  virtual size_t GetDataSize() const = 0;
+  virtual int GetDataSize() const = 0;
 
   // If there's no data in this buffer, it represents end of stream.
   bool IsEndOfStream() const;
