+++ media/base/data_source.h
@@ -30,8 +30,8 @@ class MEDIA_EXPORT DataSourceHost {
 class MEDIA_EXPORT DataSource : public base::RefCountedThreadSafe<DataSource> {
  public:
   typedef base::Callback<void(int64, int64)> StatusCallback;
-  typedef base::Callback<void(size_t)> ReadCB;
-  static const size_t kReadError;
+  typedef base::Callback<void(int)> ReadCB;
+  static const int kReadError;
 
   DataSource();
 
@@ -40,10 +40,7 @@ class MEDIA_EXPORT DataSource : public base::RefCountedThreadSafe<DataSource> {
   // Reads |size| bytes from |position| into |data|. And when the read is done
   // or failed, |read_cb| is called with the number of bytes read or
   // kReadError in case of error.
-  // TODO(hclam): should change |size| to int! It makes the code so messy
-  // with size_t and int all over the place..
-  virtual void Read(int64 position, size_t size,
-                    uint8* data,
+  virtual void Read(int64 position, int size, uint8* data,
                     const DataSource::ReadCB& read_cb) = 0;
 
   // Notifies the DataSource of a change in the current playback rate.
