+++ media/base/mock_filters.h
@@ -81,7 +81,7 @@ class MockDataSource : public DataSource {
   MOCK_METHOD0(OnAudioRendererDisabled, void());
 
   // DataSource implementation.
-  MOCK_METHOD4(Read, void(int64 position, size_t size, uint8* data,
+  MOCK_METHOD4(Read, void(int64 position, int size, uint8* data,
                           const DataSource::ReadCB& callback));
   MOCK_METHOD1(GetSize, bool(int64* size_out));
   MOCK_METHOD1(SetPreload, void(Preload preload));
